third_party/blink/web_tests/external/wpt/preload/supported-as-values_as=json_expected=1-expected.txt

This is a testharness.js-based test.
[FAIL] Test the supported value for <link rel=preload as="...">
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
